02.Agent/02.Agent.Latest-News-Events-streamlit/requirements.txt
streamlit>=1.28.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=0.1.0
langgraph>=0.2.0
langchain-tavily>=0.1.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
tavily-python>=0.3.0
